pyproject.toml
[project]
name = "fireuq"
description = "analyze fire uq data"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "absl-py",
  "numpy",
  "jaxlib",
  "jax",
  "flax",
  "h5py",
  "optax"
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf"
]

[tool.setuptools]
py-modules = ['fireuq']
